# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialApp — unified social network + messaging platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chats:
    get:
      tags: [Chats]
      operationId: getChats
      summary: Get my direct message conversations
      responses:
        '200':
          description: Chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
    post:
      tags: [Chats]
      operationId: createChat
      summary: Start or get a DM with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatInput'
      responses:
        '200':
          description: Chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
  /chats/{chatId}/messages:
    get:
      tags: [Messages]
      operationId: getChatMessages
      summary: Get messages in a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListPage'
    post:
      tags: [Messages]
      operationId: sendMessage
      summary: Send a message in a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /chats/{chatId}/read:
    post:
      tags: [Messages]
      operationId: markChatRead
      summary: Mark all messages in a chat as read
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ── Groups ────────────────────────────────────────────────────────────────
...
  /channels:
    get:
      tags: [Channels]
      operationId: getDiscoverChannels
      summary: Discover public channels
      responses:
        '200':
          description: Channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
    post:
      tags: [Channels]
      operationId: createChannel
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelInput'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
...
  /channels/{channelId}:
    get:
      tags: [Channels]
      operationId: getChannel
      summary: Get channel details
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
  /channels/{channelId}/subscribe:
    post:
      tags: [Channels]
      operationId: subscribeChannel
      summary: Subscribe to a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
    delete:
      tags: [Channels]
      operationId: unsubscribeChannel
      summary: Unsubscribe from a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /channels/{channelId}/posts:
    get:
      tags: [Channels]
      operationId: getChannelPosts
      summary: Get posts in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListPage'
    post:
      tags: [Channels]
      operationId: broadcastToChannel
      summary: Broadcast a post to a channel (admin only)
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Broadcasted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

  # ── Notifications ─────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    # ── Auth ────────────────────────────────────────────────────────────────
...
    UserSummary:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        isFollowing: { type: boolean }
...
    Post:
      type: object
      required: [id, author, content, mediaUrls, likesCount, commentsCount, isLiked, isSaved, createdAt]
      properties:
        id: { type: string }
        author:
          $ref: '#/components/schemas/UserSummary'
        content: { type: string }
        mediaUrls:
          type: array
          items: { type: string }
        mediaType: { type: ["string", "null"], enum: [image, video, "null"] }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        sharesCount: { type: integer }
        isLiked: { type: boolean }
        isSaved: { type: boolean }
        channelId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PostInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        mediaUrls:
          type: array
          items: { type: string }
        mediaType: { type: string, enum: [image, video] }
...
    PostListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        nextCursor: { type: ["string", "null"] }
...
    Chat:
      type: object
      required: [id, participant, unreadCount, createdAt]
      properties:
        id: { type: string }
        participant:
          $ref: '#/components/schemas/UserSummary'
        lastMessage:
          oneOf:
            - $ref: '#/components/schemas/Message'
            - type: 'null'
        unreadCount: { type: integer }
        createdAt: { type: string, format: date-time }
    ChatInput:
      type: object
      required: [participantId]
      properties:
        participantId: { type: string, minLength: 1 }

    # ── Messages ──────────────────────────────────────────────────────────────
    Message:
      type: object
      required: [id, senderId, content, isRead, createdAt]
      properties:
        id: { type: string }
        senderId: { type: string }
        sender:
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        content: { type: string }
        mediaUrl: { type: ["string", "null"] }
        mediaType: { type: ["string", "null"], enum: [image, video, audio, file, "null"] }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [image, video, audio, file] }
    MessageListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextCursor: { type: ["string", "null"] }

    # ── Groups ────────────────────────────────────────────────────────────────
...
    Channel:
      type: object
      required: [id, name, subscribersCount, isSubscribed, isAdmin, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        subscribersCount: { type: integer }
        isSubscribed: { type: boolean }
        isAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
    ChannelInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        avatarUrl: { type: string }

    # ── Notifications ─────────────────────────────────────────────────────────
...